Add --rbrequire (-q) option to opal cmdline tool
The short rationale behind it is that for now it's impossible to
require opal libraries like opal-browser (close to impossible - it
needs Gem paths) or opal-erubi (impossible, because it needs to
modify the compiler context to support .erubi files).

The longer rationale is that opal-rspec depends on Opal::CLI. We
can't test opal-webassembly due to that.